Add verbosity on search store
